refactor: csp.js
