+++ next.config.js
@@ -6,42 +6,12 @@ import { createRequire } from 'module';
 import CopyPlugin from 'copy-webpack-plugin';
 import { execSync } from 'child_process';
 
+import { POLICY_SETTINGS } from './csp.js';
+
 const require = createRequire(import.meta.url);
 const __dirname = fileURLToPath(dirname(import.meta.url));
 const outputPath = fileURLToPath(new URL('./public', import.meta.url));
 const polyfillsFolderPath = join(outputPath, './js/polyfills');
-
-const isDev = process.env.NODE_ENV === 'development';
-const cspConfig = {
-    'default-src': ['https:', 'wss:', "'self'"],
-    'script-src': [
-        "'self'",
-        "'unsafe-inline'",
-        "'unsafe-eval'",
-        'https://www.googletagmanager.com/',
-        'https://cdn.jsdelivr.net',
-        'https://*.vercel-scripts.com',
-        'https://*.firefly.land/',
-    ],
-    'style-src': ["'self'", "'unsafe-inline'"],
-    'worker-src': ["'self'", 'blob:'],
-};
-
-// Add Sentry DSN to CSP report-uri
-if (process.env.NEXT_PUBLIC_SENTRY_REPORT_URL) {
-    cspConfig['report-uri'] = [process.env.NEXT_PUBLIC_SENTRY_REPORT_URL];
-}
-
-if (isDev) {
-    Object.entries(cspConfig).forEach(([key, value]) => {
-        if (key === 'report-uri') return;
-        value.push('http://localhost:3000', 'ws://localhost:3000');
-    });
-}
-
-const cspHeaderContent = Object.entries(cspConfig)
-    .map(([key, value]) => `${key} ${value.join(' ')}`)
-    .join('; ');
 
 /** @type {import('next').NextConfig} */
 export default {
@@ -117,7 +87,7 @@ export default {
                     },
                     {
                         key: 'Content-Security-Policy-Report-Only',
-                        value: cspHeaderContent,
+                        value: POLICY_SETTINGS,
                     },
                 ],
             },
